fixed the login and signup navigation
