+++ src/components/Header/Header.jsx
@@ -1,13 +1,16 @@
-import { signOut } from "firebase/auth";
+import { onAuthStateChanged, signOut } from "firebase/auth";
 import netflixLogo from "../../assets/Netflix-logo.png";
 import userLogo from "../../assets/user.png";
 import { auth } from "../../utils/firebase";
 import { useNavigate } from "react-router-dom";
-import {  useSelector } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
+import { useEffect } from "react";
+import { addUser, removeUser } from "../../utils/userSlice";
 
 const Header = () => {
   const user = useSelector((store) => store.user);
   const navigate = useNavigate();
+  const dispatch = useDispatch();
   const handleSignOut = () => {
     signOut(auth)
       .then(() => {
@@ -16,9 +19,32 @@ const Header = () => {
       })
       .catch((error) => {
         // An error happened.
-        navigate('/error')
+        navigate("/error");
       });
   };
+
+  useEffect(() => {
+    onAuthStateChanged(auth, (user) => {
+      if (user) {
+        // User is signed in, see docs for a list of available properties
+        // https://firebase.google.com/docs/reference/js/auth.user
+        const { uid, email, displayName, photoURL } = user;
+        dispatch(
+          addUser({
+            uid: uid,
+            email: email,
+            displayName: displayName,
+            photoURL: photoURL,
+          })
+        );
+        navigate("/browse");
+      } else {
+        // User is signed out
+        dispatch(removeUser());
+        navigate("/");
+      }
+    });
+  }, []);
 
   return (
     <div className="flex justify-between absolute w-full px-8 py-2 bg-gradient-to-b from-black z-30">
@@ -28,7 +54,11 @@ const Header = () => {
       </button> */}
       {user && (
         <div className="flex p-5 justify-between">
-          <img className="w-12 h-12" src={user?.photoURL || userLogo} alt="userLogo" />
+          <img
+            className="w-12 h-12"
+            src={user?.photoURL || userLogo}
+            alt="userLogo"
+          />
           <button
             onClick={handleSignOut}
             className="font-bold text-white h-12 px-2"
